Removed Number Bugs

+++ app/src/main/java/com/example/offiqlresturantapp/ui/searchfood/adaptor/ListOfFoodItemToSearchAdaptor.kt
@@ -36,7 +36,7 @@ class ListOfFoodItemToSearchAdaptor(private val itemClickListerForListOfFood: It
             data.forEach {
                 str.append(it)
             }
-            return str.toString().toInt()
+            return str.toString().toFloat().toInt()
         }
     }
